added mock dao class and interface using game data that was originally in service. connected service to dao interface.
